Support armv7k architecture for watchOS 2

+++ build.sh
@@ -52,6 +52,9 @@ IPHONE_EXTENSIONS="objc"
 IPHONE_PLATFORMS="iPhoneOS iPhoneSimulator"
 IPHONE_DIR="iphone-lib"
 
+WATCHOS_PLATFORMS="WatchOS WatchSimulator"
+WATCHOS_DIR="watchos-lib"
+
 ANDROID_DIR="android-lib"
 ANDROID_PLATFORMS="arm arm-v7a arm64 mips x86 x86_64"
 
@@ -67,6 +70,7 @@ Available modes are:
     build-config-progs:                 
     build-osx:                          
     build-iphone:                       
+    build-watchos:
     build-android:                      
     build-cocoa:                        
     build-osx-framework:                
@@ -554,6 +558,36 @@ case "$MODE" in
             done
         fi
 
+        # Find watchOS SDKs
+        watchos_sdks=""
+        watchos_sdks_avail="no"
+        if [ "$xcode_home" != "none" ]; then
+            # Xcode provides the watchOS SDK
+            watchos_sdks_avail="yes"
+            for x in $WATCHOS_PLATFORMS; do
+                platform_home="$xcode_home/Platforms/$x.platform"
+                if ! [ -e "$platform_home/Info.plist" ]; then
+                    echo "Failed to find '$platform_home/Info.plist'"
+                    watchos_sdks_avail="no"
+                else
+                    sdk="$(find_iphone_sdk "$platform_home")" || exit 1
+                    if ! [ "$sdk" ]; then
+                        echo "Found no SDKs in '$platform_home'"
+                        watchos_sdks_avail="no"
+                    else
+                        if [ "$x" = "WatchSimulator" ]; then
+                            archs="i386"
+                        elif [ "$x" = "WatchOS" ]; then
+                            archs="armv7k"
+                        else
+                            continue
+                        fi
+                        word_list_append "watchos_sdks" "$x:$sdk:$archs" || exit 1
+                    fi
+                fi
+            done
+        fi
+
         # Find Android NDK
         if [ "$ANDROID_NDK_HOME" ]; then
             android_ndk_home="$ANDROID_NDK_HOME"
@@ -562,7 +596,7 @@ case "$MODE" in
         fi
 
         cat >"$CONFIG_MK" <<EOF
-REALM_VERSION       = $realm_version
+REALM_VERSION         = $realm_version
 INSTALL_PREFIX        = $install_prefix
 INSTALL_EXEC_PREFIX   = $install_exec_prefix
 INSTALL_INCLUDEDIR    = $install_includedir
@@ -578,6 +612,8 @@ ENABLE_ENCRYPTION     = $enable_encryption
 XCODE_HOME            = $xcode_home
 IPHONE_SDKS           = ${iphone_sdks:-none}
 IPHONE_SDKS_AVAIL     = $iphone_sdks_avail
+WATCHOS_SDKS          = ${watchos_sdks:-none}
+WATCHOS_SDKS_AVAIL    = $watchos_sdks_avail
 ANDROID_NDK_HOME      = $android_ndk_home
 EOF
         if ! [ "$INTERACTIVE" ]; then
@@ -603,6 +639,17 @@ EOF
                 rm -f "$IPHONE_DIR/librealm-ios.a" "$IPHONE_DIR/librealm-ios-dbg.a" || exit 1
                 rm -f "$IPHONE_DIR/realm-config" "$IPHONE_DIR/realm-config-dbg" || exit 1
                 rmdir "$IPHONE_DIR" || exit 1
+            fi
+            for x in $WATCHOS_PLATFORMS; do
+                $MAKE -C "src/realm" clean BASE_DENOM="$x" || exit 1
+            done
+            $MAKE -C "src/realm" clean BASE_DENOM="watch" || exit 1
+            if [ -e "$WATCHOS_DIR" ]; then
+                echo "Removing '$WATCHOS_DIR'"
+                rm -fr "$WATCHOS_DIR/include" || exit 1
+                rm -f "$WATCHOS_DIR/librealm-watchos.a" "$WATCHOS_DIR/librealm-watchos-dbg.a" || exit 1
+                rm -f "$WATCHOS_DIR/realm-config" "$WATCHOS_DIR/realm-config-dbg" || exit 1
+                rmdir "$WATCHOS_DIR" || exit 1
             fi
         fi
         for x in $ANDROID_PLATFORMS; do
@@ -665,22 +712,29 @@ EOF
             platform="$(printf "%s\n" "$x" | cut -d: -f1)" || exit 1
             sdk="$(printf "%s\n" "$x" | cut -d: -f2)" || exit 1
             archs="$(printf "%s\n" "$x" | cut -d: -f3 | sed 's/,/ /g')" || exit 1
-            cflags_arch="-stdlib=libc++"
+            cflags_arch="-stdlib=libc++ -fembed-bitcode"
             for y in $archs; do
                 word_list_append "cflags_arch" "-arch $y" || exit 1
             done
+            if [ "$platform" = 'iPhoneOS' ]; then
+                word_list_append "cflags_arch" "-mstrict-align" || exit 1
+                word_list_append "cflags_arch" "-miphoneos-version-min=7.0" || exit 1
+            fi
+            if [ "$platform" = 'iPhoneSimulator' ]; then
+                word_list_append "cflags_arch" "-mios-simulator-version-min=7.0" || exit 1
+            fi
             sdk_root="$xcode_home/Platforms/$platform.platform/Developer/SDKs/$sdk"
-            $MAKE -C "src/realm" "librealm-$platform.a" "librealm-$platform-dbg.a" BASE_DENOM="$platform" CFLAGS_ARCH="$cflags_arch -isysroot '$sdk_root' -mstrict-align" || exit 1
+            $MAKE -C "src/realm" "librealm-$platform.a" "librealm-$platform-dbg.a" BASE_DENOM="$platform" CFLAGS_ARCH="$cflags_arch -isysroot '$sdk_root'" || exit 1
             mkdir "$temp_dir/platforms/$platform" || exit 1
             cp "src/realm/librealm-$platform.a"     "$temp_dir/platforms/$platform/librealm.a"     || exit 1
             cp "src/realm/librealm-$platform-dbg.a" "$temp_dir/platforms/$platform/librealm-dbg.a" || exit 1
         done
-        REALM_ENABLE_FAT_BINARIES="1" $MAKE -C "src/realm" "realm-config-ios" "realm-config-ios-dbg" BASE_DENOM="ios" CFLAGS_ARCH="-DREALM_CONFIG_IOS" || exit 1
+        REALM_ENABLE_FAT_BINARIES="1" $MAKE -C "src/realm" "realm-config-ios" "realm-config-ios-dbg" BASE_DENOM="ios" CFLAGS_ARCH="-fembed-bitcode -DREALM_CONFIG_IOS" AR="libtool" ARFLAGS="-o" || exit 1
         mkdir -p "$IPHONE_DIR" || exit 1
         echo "Creating '$IPHONE_DIR/librealm-ios.a'"
-        lipo "$temp_dir/platforms"/*/"librealm.a"     -create -output "$IPHONE_DIR/librealm-ios.a"     || exit 1
+        libtool "$temp_dir/platforms"/*/"librealm.a"     -static -o "$IPHONE_DIR/librealm-ios.a"     || exit 1
         echo "Creating '$IPHONE_DIR/librealm-ios-dbg.a'"
-        lipo "$temp_dir/platforms"/*/"librealm-dbg.a" -create -output "$IPHONE_DIR/librealm-ios-dbg.a" || exit 1
+        libtool "$temp_dir/platforms"/*/"librealm-dbg.a" -static -o "$IPHONE_DIR/librealm-ios-dbg.a" || exit 1
         echo "Copying headers to '$IPHONE_DIR/include'"
         mkdir -p "$IPHONE_DIR/include" || exit 1
         cp "src/realm.hpp" "$IPHONE_DIR/include/" || exit 1
@@ -692,6 +746,61 @@ EOF
             echo "Creating '$IPHONE_DIR/$x'"
             y="$(printf "%s\n" "$x" | sed 's/realm-config/realm-config-ios/')" || exit 1
             cp "src/realm/$y" "$REALM_HOME/$IPHONE_DIR/$x" || exit 1
+        done
+        echo "Done building"
+        exit 0
+        ;;
+
+    "build-watchos")
+        auto_configure || exit 1
+        export REALM_HAVE_CONFIG="1"
+        watchos_sdks_avail="$(get_config_param "WATCHOS_SDKS_AVAIL")" || exit 1
+        if [ "$watchos_sdks_avail" != "yes" ]; then
+            echo "ERROR: Required watchOS SDKs are not available" 1>&2
+            exit 1
+        fi
+        temp_dir="$(mktemp -d /tmp/realm.build-watchos.XXXX)" || exit 1
+        mkdir "$temp_dir/platforms" || exit 1
+        xcode_home="$(get_config_param "XCODE_HOME")" || exit 1
+        watchos_sdks="$(get_config_param "WATCHOS_SDKS")" || exit 1
+        for x in $watchos_sdks; do
+            platform="$(printf "%s\n" "$x" | cut -d: -f1)" || exit 1
+            sdk="$(printf "%s\n" "$x" | cut -d: -f2)" || exit 1
+            archs="$(printf "%s\n" "$x" | cut -d: -f3 | sed 's/,/ /g')" || exit 1
+            cflags_arch="-stdlib=libc++ -fembed-bitcode"
+            for y in $archs; do
+                word_list_append "cflags_arch" "-arch $y" || exit 1
+            done
+            if [ "$platform" = 'WatchOS' ]; then
+                word_list_append "cflags_arch" "-mstrict-align" || exit 1
+                word_list_append "cflags_arch" "-mwatchos-version-min=2.0" || exit 1
+            fi
+            if [ "$platform" = 'WatchSimulator' ]; then
+                word_list_append "cflags_arch" "-mwatchos-simulator-version-min=2.0" || exit 1
+            fi
+            sdk_root="$xcode_home/Platforms/$platform.platform/Developer/SDKs/$sdk"
+            $MAKE -C "src/realm" "librealm-$platform.a" "librealm-$platform-dbg.a" BASE_DENOM="$platform" CFLAGS_ARCH="$cflags_arch -isysroot '$sdk_root'" || exit 1
+            mkdir "$temp_dir/platforms/$platform" || exit 1
+            cp "src/realm/librealm-$platform.a"     "$temp_dir/platforms/$platform/librealm.a"     || exit 1
+            cp "src/realm/librealm-$platform-dbg.a" "$temp_dir/platforms/$platform/librealm-dbg.a" || exit 1
+        done
+        REALM_ENABLE_FAT_BINARIES="1" $MAKE -C "src/realm" "realm-config-watchos" "realm-config-watchos-dbg" BASE_DENOM="watchos" CFLAGS_ARCH="-fembed-bitcode -DREALM_CONFIG_WATCHOS" AR="libtool" ARFLAGS="-o" || exit 1
+        mkdir -p "$WATCHOS_DIR" || exit 1
+        echo "Creating '$WATCHOS_DIR/librealm-watchos.a'"
+        libtool "$temp_dir/platforms"/*/"librealm.a"     -static -o "$WATCHOS_DIR/librealm-watchos.a"     || exit 1
+        echo "Creating '$WATCHOS_DIR/librealm-watchos-dbg.a'"
+        libtool "$temp_dir/platforms"/*/"librealm-dbg.a" -static -o "$WATCHOS_DIR/librealm-watchos-dbg.a" || exit 1
+        echo "Copying headers to '$WATCHOS_DIR/include'"
+        mkdir -p "$WATCHOS_DIR/include" || exit 1
+        cp "src/realm.hpp" "$WATCHOS_DIR/include/" || exit 1
+        mkdir -p "$WATCHOS_DIR/include/realm" || exit 1
+        inst_headers="$(cd "src/realm" && $MAKE --no-print-directory get-inst-headers)" || exit 1
+        (cd "src/realm" && tar czf "$temp_dir/headers.tar.gz" $inst_headers) || exit 1
+        (cd "$REALM_HOME/$WATCHOS_DIR/include/realm" && tar xzmf "$temp_dir/headers.tar.gz") || exit 1
+        for x in "realm-config" "realm-config-dbg"; do
+            echo "Creating '$WATCHOS_DIR/$x'"
+            y="$(printf "%s\n" "$x" | sed 's/realm-config/realm-config-watchos/')" || exit 1
+            cp "src/realm/$y" "$REALM_HOME/$WATCHOS_DIR/$x" || exit 1
         done
         echo "Done building"
         exit 0
@@ -810,6 +919,7 @@ EOF
 
         sh build.sh build-osx || exit 1
         sh build.sh build-iphone || exit 1
+        sh build.sh build-watchos || exit 1
 
         echo "Copying files"
         tmpdir=$(mktemp -d /tmp/$$.XXXXXX) || exit 1
@@ -819,8 +929,15 @@ EOF
         mkdir -p "$tmpdir/$BASENAME/include" || exit 1
         cp "$IPHONE_DIR/librealm-ios.a" "$tmpdir/$BASENAME" || exit 1
         cp "$IPHONE_DIR/librealm-ios-dbg.a" "$tmpdir/$BASENAME" || exit 1
+        cp "$WATCHOS_DIR/librealm-watchos.a" "$tmpdir/$BASENAME" || exit 1
+        cp "$WATCHOS_DIR/librealm-watchos-dbg.a" "$tmpdir/$BASENAME" || exit 1
         cp -r "$IPHONE_DIR/include/"* "$tmpdir/$BASENAME/include" || exit 1
         for x in $iphone_sdks; do
+            platform="$(printf "%s\n" "$x" | cut -d: -f1)" || exit 1
+            cp "src/realm/librealm-$platform.a" "$tmpdir/$BASENAME" || exit 1
+            cp "src/realm/librealm-$platform-dbg.a" "$tmpdir/$BASENAME" || exit 1
+        done
+        for x in $watchos_sdks; do
             platform="$(printf "%s\n" "$x" | cut -d: -f1)" || exit 1
             cp "src/realm/librealm-$platform.a" "$tmpdir/$BASENAME" || exit 1
             cp "src/realm/librealm-$platform-dbg.a" "$tmpdir/$BASENAME" || exit 1
